+++ app/components/base/quick-list/index.ts
@@ -1,4 +1,5 @@
 export * from "./quick-list.component";
-export * from "./quick-list-item.component";
+export * from "./quick-list-item";
 export * from "./quick-list.module";
+export * from "./quick-list-item-status";
 export * from "./no-item.component";
