+++ app/components/base/quick-list/quick-list-item-status/index.ts
@@ -1,0 +1,1 @@
+export * from "./quick-list-item-status.component";
