+++ app/components/base/quick-list/quick-list-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./quick-list-item.component";
