+++ app/components/base/quick-list/quick-list-item/quick-list-item.scss
