+++ app/components/base/quick-list/quick-list.module.ts
@@ -7,22 +7,20 @@ import { MaterialModule } from "app/core";
 import { VirtualScrollModule } from "app/components/base/virtual-scroll";
 import { ContextMenuModule } from "../context-menu";
 import { NoItemComponent } from "./no-item.component";
-import { QuickListItemComponent } from "./quick-list-item.component";
-import { QuickListComponent, QuickListItemStatusComponent } from "./quick-list.component";
+import { QuickListItemComponent } from "./quick-list-item";
+import { QuickListItemStatusComponent } from "./quick-list-item-status";
+import { QuickListComponent } from "./quick-list.component";
+
+const publicComponents = [
+    NoItemComponent,
+    QuickListComponent,
+    QuickListItemComponent,
+    QuickListItemStatusComponent,
+];
 
 @NgModule({
-    declarations: [
-        NoItemComponent,
-        QuickListComponent,
-        QuickListItemComponent,
-        QuickListItemStatusComponent,
-    ],
-    exports: [
-        NoItemComponent,
-        QuickListComponent,
-        QuickListItemComponent,
-        QuickListItemStatusComponent,
-    ],
+    declarations: publicComponents,
+    exports: publicComponents,
     imports: [
         BrowserModule,
         FormsModule,
